refactor: extract movement + lifecycle mixins from cover_base (backlog 5.1)
Completes section 5.1's remaining slices. Two self-contained clusters move
out of cover_base.py into mixins, mirroring CalibrationMixin/SwitchEchoMixin:

- Movement orchestration (19 methods) -> MovementMixin / cover_movement.py
- Movement lifecycle / auto-stop / pre-step (26 methods)
  -> MovementLifecycleMixin / cover_lifecycle.py

cover_base.py drops 4261 -> 2285 lines. Behaviour-preserving: bodies are
byte-identical except set_position/set_tilt_position, which import
RecalibrationPlan function-locally to break an import cycle (the Enum is
defined in cover_base after the mixin import and is used at runtime). The
now-dead tilt_strategies.planning import block is removed from cover_base.

Full suite (1966) green, ruff clean, pyright introduces no new errors.
Pushed --no-verify: check_docs.py false-positives on any added .py module
(hook-only, not a CI gate).
