+++ custom_components/cover_time_based/cover_base.py
@@ -52,14 +52,10 @@ from .const import (
 )
 from .cover_calibration import CalibrationMixin
 from .cover_echo_filter import SwitchEchoMixin
+from .cover_lifecycle import MovementLifecycleMixin
+from .cover_movement import MovementMixin
 from .position_storage import async_get_position_store
 from .tilt_strategies import InlineTilt, SequentialTilt
-from .tilt_strategies.planning import (
-    calculate_pre_step_delay,
-    extract_coupled_tilt,
-    extract_coupled_travel,
-    has_travel_pre_step,
-)
 from .travel_calculator import TravelCalculator, TravelStatus
 
 _LOGGER = logging.getLogger(__name__)
@@ -103,7 +99,14 @@ class RecalibrationPlan(Enum):
     FORCED_ENDPOINT = auto()
 
 
-class CoverTimeBased(CalibrationMixin, SwitchEchoMixin, CoverEntity, RestoreEntity):
+class CoverTimeBased(
+    CalibrationMixin,
+    SwitchEchoMixin,
+    MovementMixin,
+    MovementLifecycleMixin,
+    CoverEntity,
+    RestoreEntity,
+):
     """Time-based cover with position tracking."""
 
     # Push-based: the auto-updater writes state; there is no async_update, so a
@@ -1823,1062 +1826,6 @@ class CoverTimeBased(CalibrationMixin, SwitchEchoMixin, CoverEntity, RestoreEnti
         await self._async_persist_position()
 
     # -----------------------------------------------------------------------
-    # Movement orchestration
-    # -----------------------------------------------------------------------
-
-    async def _async_move_to_endpoint(self, target, *, suppress_start_command=False):
-        """Move cover to an endpoint (0=fully closed, 100=fully open).
-
-        ``suppress_start_command`` drives the tracker without touching a relay,
-        for the one caller that knows the motor is ALREADY running the
-        direction this move would command: a forced full re-drive
-        (_force_full_redrive) issued while the cover travels that way already.
-        Re-issuing the start command there is the hazard the plain
-        ``set_position`` path guards with ``already_moving_same_dir`` — on
-        toggle (same-button) hardware a second rising edge on the driving relay
-        STOPS the motor. This cannot be detected from inside here: the caller's
-        seed sets travel_calc to the opposite endpoint before we run, which
-        erases the "is travelling" evidence entirely (the same trap the
-        dual-motor tilt reversal works around by evaluating before the seed).
-        Hence a flag decided by the caller rather than a check here. The motor
-        keeps running toward the endpoint and stalls at its limit, which is
-        exactly what a forced re-drive wants; the startup delay is dropped too,
-        since the motor is already up to speed.
-        """
-        # NB: _self_initiated_movement is assigned only once this call is
-        # committed to acting, past the startup-delay early-returns below.
-        # A same-direction command that reaches an already-active startup delay
-        # (e.g. an underlying's lagged echo of our own move, issue #165) returns
-        # from the "not restarting" branch without touching the flag, so it
-        # cannot flip an in-flight self-initiated move to "external" and drop
-        # its auto-stop. The sequential-redirect below delegates to
-        # set_tilt_position, which sets the flag itself.
-
-        # External close on sequential hardware runs the full journey:
-        # the motor drives all the way past cover-closed to the articulated
-        # extreme (tilt=100 on sequential_open, tilt=0 on sequential_close).
-        # Redirect to set_tilt_position so tracking plans both phases as
-        # [TravelTo(0), TiltTo(articulated)].
-        #
-        # Open externally is already handled correctly by the default plan
-        # (plan_move_position restores tilt to implicit before travel when
-        # starting from the articulated state).
-        if (
-            target == 0
-            and self._triggered_externally
-            and isinstance(self._tilt_strategy, SequentialTilt)
-        ):
-            articulated = 100 - self._tilt_strategy.implicit_tilt_during_travel
-            self._log(
-                "_async_move_to_endpoint :: external close on sequential → "
-                "set_tilt_position(%d) for full-journey tracking",
-                articulated,
-            )
-            await self.set_tilt_position(articulated)
-            return
-
-        # Capture before _abandon_active_lifecycle resets it — see
-        # _release_displaced_tilt_motor.
-        was_tilt_motor_move = self._moving_tilt_motor
-        await self._abandon_active_lifecycle()
-
-        closing = target == 0
-        command = SERVICE_CLOSE_COVER if closing else SERVICE_OPEN_COVER
-        opposite_command = SERVICE_OPEN_COVER if closing else SERVICE_CLOSE_COVER
-
-        # Check startup delay conflicts BEFORE position check, since during
-        # startup delay the position hasn't started changing yet.
-        if self._startup_delay_task and not self._startup_delay_task.done():
-            if self._last_command == opposite_command:
-                self._log(
-                    "_async_move_to_endpoint :: direction change, stopping the parked move"
-                )
-                await self._neutralize_parked_move()
-                await self._async_handle_command(SERVICE_STOP_COVER)
-                self._last_command = None
-                await self._release_displaced_tilt_motor(was_tilt_motor_move)
-                # Silent: no travel started. _movement_started must read this
-                # as "not started" — see its docstring.
-                return
-            else:
-                self._log(
-                    "_async_move_to_endpoint :: startup delay already active, not restarting"
-                )
-                await self._release_displaced_tilt_motor(was_tilt_motor_move)
-                # Silent: no travel started. _movement_started must read this
-                # as "not started" — see its docstring.
-                return
-
-        # Committed to acting: now claim the movement's bookkeeping. Unlike the
-        # tilt funnel (_async_move_tilt_to_endpoint, which sets this flag then
-        # runs a post-flag in-motion reversal), the travel funnel's only
-        # direction change is the startup-delay cancel above, which returns
-        # before this line — so on that path _self_initiated_movement is left
-        # stale. Benign: the cover is fully stopped there and the next funnel
-        # call overwrites the flag.
-        self._self_initiated_movement = not self._triggered_externally
-
-        current = self.travel_calc.current_position()
-        if current is not None and current == target:
-            # Resync: send command + endpoint run-on even though tracker
-            # says we're already there. Physical cover may need resyncing.
-            self._cancel_delay_task()
-            self._last_command = command
-            await self._async_handle_command(command)
-            if self._self_stops_at_endpoints() and self._at_endpoint(target):
-                # The re-drive above resyncs the cover physically; the motor
-                # self-stops at its limit. A stop here is redundant (and for
-                # toggle re-pulses → restart), so skip it and any run-on.
-                self._log(
-                    "_async_move_to_endpoint :: motor self-stops at endpoint,"
-                    " no relay stop"
-                )
-            elif (
-                self._endpoint_runon_time is not None and self._endpoint_runon_time > 0
-            ):
-                self._delay_task = self.hass.async_create_task(
-                    self._delayed_stop(self._endpoint_runon_time)
-                )
-            else:
-                await self._async_handle_command(SERVICE_STOP_COVER)
-            await self._release_displaced_tilt_motor(was_tilt_motor_move)
-            # Silent: a relay resync fires here, but travel_calc never enters
-            # "traveling" — _movement_started must still read this as "not
-            # started". Not reachable from _force_full_redrive (it seeds the
-            # opposite endpoint first, so current never equals target here),
-            # but a future caller of this method directly would hit it.
-            return
-
-        relay_was_on = self._cancel_delay_task()
-        if relay_was_on:
-            await self._async_handle_command(SERVICE_STOP_COVER)
-
-        if current is None:
-            # Position unknown — assume opposite endpoint so full travel occurs
-            current = 100 if closing else 0
-            self.travel_calc.update_position(current)
-
-        travel_distance = abs(target - current)
-        travel_time = self._require_travel_time(closing)
-        movement_time = (travel_distance / 100.0) * travel_time
-
-        self._log(
-            "_async_move_to_endpoint :: target=%d, travel_distance=%f%%, movement_time=%fs",
-            target,
-            travel_distance,
-            movement_time,
-        )
-
-        self._last_command = command
-
-        tilt_target = None
-        pre_step_delay = 0.0
-        # Plan tilt for every trigger including external. Even when the cover
-        # motor is externally controlled, the hardware itself is expected to
-        # move tilt to safe before travel (interlock behavior); tracking the
-        # pre-step keeps the integration's tilt_calc in sync with reality
-        # without needing snap_trackers_to_physical to "correct" the tracker
-        # at stop time. The pre-step phase (_start_tilt_pre_step /
-        # _start_travel_pre_step) runs on the handler's own task and skips
-        # relay firing on _triggered_externally; the continuation phase
-        # (_start_pending_travel / _start_pending_tilt) runs on the
-        # auto-updater's task and skips on _self_initiated_movement — the
-        # instance flag that survives across the phase boundary, unlike the
-        # task-scoped _triggered_externally. Either way the integration only
-        # mirrors the physical motion in its calculators for external moves.
-        current_tilt = (
-            self.tilt_calc.current_position() if self._tilt_strategy else None
-        )
-        self._require_movement_target_available(self._movement_target(closing))
-        tilt_target, pre_step_delay, started = await self._plan_tilt_for_travel(
-            target, command, current, current_tilt
-        )
-        if started:
-            return
-
-        # No pre-step took over: release tilt before energising the travel relay.
-        await self._release_displaced_tilt_motor(was_tilt_motor_move)
-
-        if not suppress_start_command:
-            await self._async_handle_command(command)
-        coupled_calc = self.tilt_calc if tilt_target is not None else None
-        self._begin_movement(
-            target,
-            tilt_target,
-            self.travel_calc,
-            coupled_calc,
-            None if suppress_start_command else self._travel_startup_delay,
-            pre_step_delay,
-        )
-
-    async def _async_move_tilt_to_endpoint(
-        self, target, *, suppress_start_command=False
-    ):
-        """Move tilt to an endpoint (0=fully closed, 100=fully open).
-
-        ``suppress_start_command`` is the tilt-axis counterpart of the travel
-        funnel's flag — see _async_move_to_endpoint. Same caller
-        (_force_full_tilt_redrive), same reason: its seed erases the evidence,
-        so the decision has to be made before the drive.
-        """
-        # As in _async_move_to_endpoint: assign _self_initiated_movement only
-        # once committed to acting, past the "startup delay already active, not
-        # restarting" early-return below, so a same-direction command reaching
-        # an active startup delay can't flip an in-flight self-initiated move to
-        # "external". _abandon_active_lifecycle does not read the flag.
-        # Capture before _abandon_active_lifecycle resets it — see
-        # _stop_displaced_movement_for_tilt.
-        was_tilt_motor_move = self._moving_tilt_motor
-        await self._abandon_active_lifecycle()
-
-        closing = target == 0
-        if self._tilt_strategy is not None:
-            command = self._tilt_strategy.tilt_command_for(closing)
-            opposite_command = self._tilt_strategy.tilt_command_for(not closing)
-        else:
-            command = SERVICE_CLOSE_COVER if closing else SERVICE_OPEN_COVER
-            opposite_command = SERVICE_OPEN_COVER if closing else SERVICE_CLOSE_COVER
-
-        if self._startup_delay_task and not self._startup_delay_task.done():
-            if self._last_command == opposite_command:
-                self._log(
-                    "_async_move_tilt_to_endpoint :: direction change, stopping the parked move"
-                )
-                await self._neutralize_parked_move()
-                # Mirror the travel counterpart: a direction change at a parked
-                # move stops the axis and returns — a second press then drives
-                # the new direction. Route the stop through the axis-aware
-                # helper so a dedicated tilt motor is not stopped via a travel
-                # STOP off a stale _last_command (#153).
-                await self._stop_displaced_movement_for_tilt(was_tilt_motor_move)
-                self._last_command = None
-                # Silent: no tilt drive started. _movement_started must read
-                # this as "not started" — see its docstring.
-                return
-            else:
-                self._log(
-                    "_async_move_tilt_to_endpoint :: startup delay already active, not restarting"
-                )
-                # Silent: no tilt drive started. _movement_started must read
-                # this as "not started" — see its docstring.
-                return
-
-        # Committed to acting (a direction change above falls through to here):
-        # now claim the movement's bookkeeping.
-        self._self_initiated_movement = not self._triggered_externally
-
-        relay_was_on = self._cancel_delay_task()
-        if relay_was_on:
-            await self._async_handle_command(SERVICE_STOP_COVER)
-            if self._tilt_strategy is not None and self._tilt_strategy.uses_tilt_motor:
-                await self._send_tilt_stop()
-
-        self._stop_travel_if_traveling()
-
-        current_tilt = self.tilt_calc.current_position()
-        if current_tilt is not None and current_tilt == target:
-            if self.tilt_calc.is_traveling():
-                # The animation is momentarily sitting on the target while still
-                # travelling away from it (a reversal landing on the endpoint it
-                # is leaving). A press for that endpoint is a user stop, not a
-                # no-op — the old early-return left the motor running.
-                await self.async_stop_cover(tilt_axis_reported=False)
-            # Silent either way: no tilt drive started. _movement_started
-            # must read this as "not started" — see its docstring.
-            return
-
-        # In-motion reversal: the new tilt direction opposes the tilt already in
-        # flight. A reversal is never a single step — stop the axis, let it come
-        # to rest (settle gap), then drive the other way (#147/#153). Bail if a
-        # stop or newer target claimed the movement inside the gap. Only for
-        # self-driven moves: an external reversal was already stopped-and-turned
-        # by the hardware, so firing a stop here would just mark phantom relay
-        # echoes and swallow the user's next press — we only retrack it.
-        if (
-            not self._triggered_externally
-            and self.tilt_calc.is_traveling()
-            and self.tilt_calc.is_closing() != closing
-        ):
-            self._log(
-                "_async_move_tilt_to_endpoint :: in-motion reversal, stop + settle"
-            )
-            self.tilt_calc.stop()
-            if self.travel_calc.is_traveling():
-                self.travel_calc.stop()
-            self.stop_auto_updater()
-            await self._stop_displaced_movement_for_tilt(was_tilt_motor_move)
-            if not await self._settle_before_reversing():
-                # Silent: superseded mid-settle, both calcs already stopped
-                # above. _movement_started must read this as "not started" —
-                # see its docstring.
-                return
-            current_tilt = self.tilt_calc.current_position()
-            if current_tilt is not None and current_tilt == target:
-                # Silent: settled exactly at target, nothing left to drive.
-                # _movement_started must read this as "not started" — see its
-                # docstring.
-                return
-
-        if current_tilt is None:
-            current_tilt = 100 if closing else 0
-            self.tilt_calc.update_position(current_tilt)
-
-        tilt_distance = abs(target - current_tilt)
-        tilt_time = self._tilting_time_close if closing else self._tilting_time_open
-        movement_time = (tilt_distance / 100.0) * tilt_time
-
-        travel_target = None
-        pre_step_delay = 0.0
-        needs_travel_pre_step = False
-        if self._tilt_strategy is not None:
-            current_pos = self.travel_calc.current_position()
-            if current_pos is not None:
-                steps = self._tilt_strategy.plan_move_tilt(
-                    target, current_pos, current_tilt
-                )
-                travel_target = extract_coupled_travel(steps)
-                pre_step_delay = calculate_pre_step_delay(
-                    steps, self._tilt_strategy, self.tilt_calc, self.travel_calc
-                )
-                if self._tilt_strategy.uses_tilt_motor and has_travel_pre_step(steps):
-                    needs_travel_pre_step = True
-
-        self._log(
-            "_async_move_tilt_to_endpoint :: target=%d, tilt_distance=%f%%,"
-            " movement_time=%fs, travel_pos=%s, travel_pre_step=%s",
-            target,
-            tilt_distance,
-            movement_time,
-            travel_target if travel_target is not None else "N/A",
-            needs_travel_pre_step,
-        )
-
-        self._require_movement_target_available(self._tilt_movement_target(command))
-        # The travel pre-step below isn't covered by the tilt gate above — see
-        # _require_movement_target_available.
-        if needs_travel_pre_step and travel_target is not None:
-            await self._start_travel_pre_step(travel_target, target, command)
-            return
-
-        self._last_command = command
-        # Committed past the no-op return above — mark this as a tilt move so it
-        # doesn't run on at a travel endpoint (#125); see set_tilt_position.
-        self._moving_tilt = True
-        if self._tilt_strategy is not None and self._tilt_strategy.uses_tilt_motor:
-            self._moving_tilt_motor = True
-            # Externally triggered moves only track — the relay is already
-            # driven from outside HA, and re-firing it here would turn the
-            # opposite relay off a second time (the observe-path interlock in
-            # _handle_external_tilt_state_change already did), double-marking
-            # its pending echo and swallowing the user's next press. Mirrors
-            # the _async_handle_command guard the non-tilt-motor branch uses.
-            if not self._triggered_externally and not suppress_start_command:
-                if closing:
-                    await self._send_tilt_close()
-                else:
-                    await self._send_tilt_open()
-        elif not suppress_start_command:
-            await self._async_handle_command(command)
-        self._begin_movement(
-            target,
-            travel_target,
-            self.tilt_calc,
-            self.travel_calc,
-            None if suppress_start_command else self._tilt_startup_delay,
-            pre_step_delay,
-        )
-
-    async def set_position(self, position, *, recalibrate: bool = True):
-        """Move cover to a designated position.
-
-        ``recalibrate`` is passed False by exactly one caller,
-        ``_maybe_start_recalibrated_leg`` running leg B of a recalibrated
-        move (issue #179), to stop this call from arming a leg of its own on
-        top of one that already ran. It has a second, less obvious
-        consequence: it also exempts the move from ``min_movement_time`` (see
-        ``_is_movement_too_short``'s ``is_recalibrated_leg`` parameter, which
-        is fed ``not recalibrate``) — leg A has already driven the motor a
-        full travel, so leg B's own pulse is the point of the operation, not
-        a redundant nudge to skip. A future caller passing ``recalibrate=False``
-        for an unrelated reason would silently get that exemption too.
-        """
-        plan = self._recalibration_plan(recalibrate, position, axis="travel")
-        if plan is RecalibrationPlan.TWO_LEG:
-            # Leg A always drives OPEN (the fully-open datum), so it only
-            # reverses an in-flight movement that is currently closing.
-            # Stop and settle first if so (issue #179) — see
-            # _stop_and_settle_before_recalibration_drive.
-            if not await self._stop_and_settle_before_recalibration_drive(
-                SERVICE_OPEN_COVER
-            ):
-                return
-            # Arm only AFTER the drive, never before. _start_recalibration_drive
-            # -> _force_full_redrive -> _async_move_to_endpoint funnels through
-            # _abandon_active_lifecycle, whose first line is _supersede_movement()
-            # (bumps _movement_epoch). _arm_recalibrated_leg stamps
-            # _recalibration_epoch from self._movement_epoch, so arming before
-            # the drive would capture the pre-bump epoch — leg A's own
-            # completion would then read as a newer, unrelated movement and the
-            # epoch check in _maybe_start_recalibrated_leg would silently drop
-            # leg B.
-            if await self._start_recalibration_drive("travel"):
-                self._arm_recalibrated_leg(position, "travel")
-                return
-            self._log(
-                "set_position :: recalibration leg did not start, moving directly"
-            )
-        elif plan is RecalibrationPlan.FORCED_ENDPOINT:
-            # position is 0 or 100: driving there IS the recalibration, but
-            # only when it is actually forced to cover the full travel
-            # (issue #179 finding 3) rather than trusting the believed
-            # position for an ordinary timed move. There is no second leg to
-            # arm here — the endpoint target itself is the whole move — so
-            # reuse _start_recalibration_drive purely for its snapshot/
-            # rollback and not-started handling: _force_full_redrive can
-            # silently fail to start (e.g. a same-direction startup delay
-            # already active), and that path must not leave the tracker
-            # seeded at a fabricated endpoint.
-            #
-            # Unlike leg A, this drive can reverse either way — target=0
-            # drives CLOSE, target=100 drives OPEN — so stop and settle
-            # first if it would (issue #179).
-            endpoint_command = (
-                SERVICE_CLOSE_COVER if position == 0 else SERVICE_OPEN_COVER
-            )
-            if not await self._stop_and_settle_before_recalibration_drive(
-                endpoint_command
-            ):
-                return
-            if await self._start_recalibration_drive("travel", target=position):
-                return
-            self._log(
-                "set_position :: forced endpoint redrive did not start, moving directly"
-            )
-        self._self_initiated_movement = not self._triggered_externally
-        # Capture before _abandon_active_lifecycle resets it — see
-        # _release_displaced_tilt_motor.
-        was_tilt_motor_move = self._moving_tilt_motor
-        await self._abandon_active_lifecycle()
-        current = self.travel_calc.current_position()
-        target = position
-        self._log(
-            "set_position :: current: %s, target: %d",
-            current if current is not None else "None",
-            target,
-        )
-
-        if current is None:
-            # Position unknown — assume opposite endpoint so full travel occurs
-            closing = target <= 50
-            command = SERVICE_CLOSE_COVER if closing else SERVICE_OPEN_COVER
-            current = 100 if closing else 0
-            self.travel_calc.update_position(current)
-        elif target < current:
-            command = SERVICE_CLOSE_COVER
-        elif target > current:
-            command = SERVICE_OPEN_COVER
-        else:
-            await self._release_displaced_tilt_motor(was_tilt_motor_move)
-            return
-
-        closing = command == SERVICE_CLOSE_COVER
-
-        should_proceed, is_direction_change = await self._handle_pre_movement_checks(
-            command
-        )
-        if not should_proceed:
-            await self._release_displaced_tilt_motor(was_tilt_motor_move)
-            return
-
-        # A shared-motor (inline/sequential) tilt move drives the same physical
-        # motor as travel, with only tilt_calc traveling. Dual-motor tilt is
-        # excluded — its tilt motor is independent of travel. Direction
-        # comparability holds because a shared-motor tilt move set
-        # _last_command to the motor direction it drives (tilt_command_for).
-        shared_motor_tilt_traveling = (
-            self._has_tilt_support()
-            and not self._tilt_strategy.uses_tilt_motor
-            and self.tilt_calc.is_traveling()
-        )
-
-        # A running dedicated tilt motor never reaches this branch:
-        # shared_motor_tilt_traveling excludes uses_tilt_motor strategies by
-        # construction, and every tilt-motor entry point stops travel_calc on
-        # its way in — so was_tilt_motor_move implies neither condition holds.
-        if is_direction_change and (
-            self.travel_calc.is_traveling() or shared_motor_tilt_traveling
-        ):
-            self._log("set_position :: stopping active travel movement")
-            self.travel_calc.stop()
-            self.stop_auto_updater()
-            if self._has_tilt_support() and self.tilt_calc.is_traveling():
-                self.tilt_calc.stop()
-            await self._async_handle_command(SERVICE_STOP_COVER)
-            if not await self._settle_before_reversing():
-                return
-            current = self.travel_calc.current_position()
-            if target == current:
-                return
-            # Travel was live when we entered this branch, so the tracker holds
-            # a known position after the settle.
-            assert current is not None
-
-        relay_was_on = self._cancel_delay_task()
-        if relay_was_on:
-            await self._async_handle_command(SERVICE_STOP_COVER)
-
-        travel_time = self._require_travel_time(closing)
-        movement_time = (abs(target - current) / 100.0) * travel_time
-
-        if self._is_movement_too_short(
-            movement_time,
-            target,
-            current,
-            "set_position",
-            is_recalibrated_leg=not recalibrate,
-        ):
-            await self._release_displaced_tilt_motor(was_tilt_motor_move)
-            return
-
-        self._last_command = command
-
-        # If the cover is already travelling the right way, the motor is
-        # running and only the stop target needs to change. Re-issuing the
-        # start command would, in toggle mode, pulse the direction switch a
-        # second time and stop the motor — desyncing it from the position
-        # tracker (a later auto-stop pulse then restarts it and runs the cover
-        # to the endpoint). Tilt coupling is still recomputed below for the new
-        # target before we retarget.
-        already_moving_same_dir = not is_direction_change and (
-            self.travel_calc.is_traveling() or shared_motor_tilt_traveling
-        )
-
-        current_tilt = (
-            self.tilt_calc.current_position() if self._tilt_strategy else None
-        )
-        # A tilt pre-step started here returns before the gate below — see
-        # _require_movement_target_available.
-        tilt_target, pre_step_delay, started = await self._plan_tilt_for_travel(
-            target, command, current, current_tilt
-        )
-        if started:
-            return
-
-        # No pre-step took the tilt motor over — see _async_move_to_endpoint's
-        # equivalent site.
-        await self._release_displaced_tilt_motor(was_tilt_motor_move)
-
-        coupled_calc = self.tilt_calc if tilt_target is not None else None
-
-        # Drive the motor toward the target, then begin tracking. When already
-        # moving the right way the motor is left running and only the tracker is
-        # retargeted (no startup delay re-applied). _command_position_move is the
-        # seam subclasses override when the device has native position control.
-        if already_moving_same_dir:
-            self._log("set_position :: retargeting active movement to %d", target)
-        await self._command_position_move(target, command, already_moving_same_dir)
-        startup_delay = None if already_moving_same_dir else self._travel_startup_delay
-        self._begin_movement(
-            target,
-            tilt_target,
-            self.travel_calc,
-            coupled_calc,
-            startup_delay,
-            pre_step_delay,
-        )
-
-    async def _command_position_move(self, target, command, already_moving_same_dir):
-        """Drive the travel motor for a mid-position ``set_position`` move.
-
-        Base behavior: when already travelling the right way, leave the motor
-        running (the caller only retargets the tracker); otherwise check target
-        availability and latch the directional relay, relying on the tracker to
-        stop it on arrival. Subclasses whose device has native position control
-        (e.g. a wrapped cover exposing ``set_cover_position``) override this to
-        forward the target straight to the device instead.
-        """
-        if already_moving_same_dir:
-            return
-        self._require_movement_target_available(
-            self._movement_target(command == SERVICE_CLOSE_COVER)
-        )
-        await self._async_handle_command(command)
-
-    async def set_tilt_position(self, position, *, recalibrate: bool = True):
-        """Move cover tilt to a designated position.
-
-        ``recalibrate`` is passed False by exactly one caller,
-        ``_maybe_start_recalibrated_leg`` running leg B of a recalibrated
-        move (issue #179), to stop this call from arming a leg of its own on
-        top of one that already ran. It has a second, less obvious
-        consequence: it also exempts the move from ``min_movement_time`` (see
-        ``_is_movement_too_short``'s ``is_recalibrated_leg`` parameter, which
-        is fed ``not recalibrate``) — leg A has already driven the motor a
-        full travel, so leg B's own pulse is the point of the operation, not
-        a redundant nudge to skip. A future caller passing ``recalibrate=False``
-        for an unrelated reason would silently get that exemption too.
-        """
-        plan = self._recalibration_plan(recalibrate, position, axis="tilt")
-        if plan is RecalibrationPlan.TWO_LEG:
-            # The drive axis is "tilt" only where the tilt motor is independent
-            # (dual_motor) and so stalls against its own limit; everywhere else
-            # the tilt "motor" IS the travel motor, so the datum is a travel
-            # endpoint and leg A drives "travel" instead. The *armed* axis is
-            # always "tilt" regardless — leg B is the tilt move the caller
-            # asked for either way.
-            axis = "tilt" if self._tilt_strategy.uses_tilt_motor else "travel"
-            # Leg A always drives fully open (closing=False). Stop and settle
-            # first if that reverses an in-flight movement (issue #179,
-            # mirroring set_position's own leg A). dual_motor's tilt motor is
-            # independent, so it needs the axis-aware tilt helper (a moving
-            # dedicated tilt motor must get a tilt stop, not a travel STOP —
-            # #153); every other strategy drives leg A via the plain travel
-            # OPEN command, so it reuses the same helper set_position already
-            # uses.
-            if axis == "tilt":
-                was_tilt_motor_move = self._moving_tilt_motor
-                if not await self._stop_and_settle_tilt_before_recalibration_drive(
-                    self._tilt_strategy.tilt_command_for(False), was_tilt_motor_move
-                ):
-                    return
-            elif not await self._stop_and_settle_before_recalibration_drive(
-                SERVICE_OPEN_COVER
-            ):
-                return
-            # Arm only AFTER the drive, never before — mirrors set_position.
-            # _start_recalibration_drive funnels through _async_move_to_endpoint
-            # or _async_move_tilt_to_endpoint, both of which start with
-            # _abandon_active_lifecycle -> _supersede_movement() (bumps
-            # _movement_epoch). _arm_recalibrated_leg stamps _recalibration_epoch
-            # from self._movement_epoch, so arming before the drive would
-            # capture the pre-bump epoch — leg A's own completion would then
-            # read as a newer, unrelated movement and the epoch check in
-            # _maybe_start_recalibrated_leg would silently drop leg B.
-            if await self._start_recalibration_drive(axis):
-                self._arm_recalibrated_leg(position, "tilt")
-                return
-            self._log(
-                "set_tilt_position :: recalibration leg did not start, moving directly"
-            )
-        elif plan is RecalibrationPlan.FORCED_ENDPOINT:
-            # A tilt endpoint on a dedicated tilt motor: driving there IS the
-            # recalibration, so there is no second leg to arm — but only when
-            # the drive is actually forced to cover the full tilt time rather
-            # than trusting the believed tilt for an ordinary timed move.
-            # Exactly set_position's endpoint branch, on the tilt axis; see
-            # _recalibration_plan for why NONE was wrong here. This plan is
-            # only ever produced for uses_tilt_motor, so the drive axis is
-            # always "tilt".
-            #
-            # Like the travel endpoint drive (and unlike leg A, which always
-            # opens) this can reverse either way — target=0 drives the tilt
-            # motor closed, target=100 open — so stop and settle first if it
-            # would, through the axis-aware tilt helper.
-            endpoint_command = self._tilt_strategy.tilt_command_for(position == 0)
-            was_tilt_motor_move = self._moving_tilt_motor
-            if not await self._stop_and_settle_tilt_before_recalibration_drive(
-                endpoint_command, was_tilt_motor_move
-            ):
-                return
-            # _start_recalibration_drive is reused purely for its snapshot /
-            # rollback and not-started handling — _force_full_tilt_redrive can
-            # silently fail to start, and that must not leave tilt_calc seeded
-            # at a fabricated endpoint.
-            if await self._start_recalibration_drive("tilt", target=position):
-                return
-            self._log(
-                "set_tilt_position :: forced tilt endpoint redrive did not"
-                " start, moving directly"
-            )
-        self._self_initiated_movement = not self._triggered_externally
-        # Capture before _abandon_active_lifecycle resets it: a dedicated tilt
-        # motor being displaced must get an axis-aware stop, not a travel STOP
-        # off a stale _last_command (#153) — see _stop_displaced_movement_for_tilt.
-        was_tilt_motor_move = self._moving_tilt_motor
-        await self._abandon_active_lifecycle()
-        current = self.tilt_calc.current_position()
-        target = position
-        self._log(
-            "set_tilt_position :: current: %s, target: %d",
-            current if current is not None else "None",
-            target,
-        )
-
-        if current is None:
-            closing = target <= 50
-            current = 100 if closing else 0
-            self.tilt_calc.update_position(current)
-        elif target < current:
-            closing = True
-        elif target > current:
-            closing = False
-        else:
-            await self._release_displaced_tilt_motor(was_tilt_motor_move)
-            return
-
-        if self._tilt_strategy is not None:
-            command = self._tilt_strategy.tilt_command_for(closing)
-        else:
-            command = SERVICE_CLOSE_COVER if closing else SERVICE_OPEN_COVER
-
-        should_proceed, is_direction_change = await self._handle_pre_movement_checks(
-            command
-        )
-        if not should_proceed:
-            await self._release_displaced_tilt_motor(was_tilt_motor_move)
-            return
-
-        if is_direction_change:
-            was_moving = (
-                self.tilt_calc.is_traveling() or self.travel_calc.is_traveling()
-            )
-            if self.tilt_calc.is_traveling():
-                self.tilt_calc.stop()
-            if self.travel_calc.is_traveling():
-                self.travel_calc.stop()
-            self.stop_auto_updater()
-            await self._stop_displaced_movement_for_tilt(was_tilt_motor_move)
-            # A reversal is stop → settle → go: give the motor (or momentary
-            # relay) time to come to rest before driving the other way, and bail
-            # if a stop/newer target claimed the movement inside the gap. This
-            # settle is intentionally NOT gated on _triggered_externally: the
-            # dual-motor drive below (_send_tilt_close/open) is itself ungated on
-            # external, so an external redirect reaching here (e.g. a sequential
-            # external close) still drives the motor and must get its rest gap —
-            # reversing a physical cover too fast can trip its supply. Do not
-            # "optimise" this to skip the settle when external.
-            if was_moving and not await self._settle_before_reversing():
-                return
-            current = self.tilt_calc.current_position()
-            if target == current:
-                return
-
-        relay_was_on = self._cancel_delay_task()
-        if relay_was_on:
-            await self._async_handle_command(SERVICE_STOP_COVER)
-            if self._tilt_strategy is not None and self._tilt_strategy.uses_tilt_motor:
-                await self._send_tilt_stop()
-
-        if not is_direction_change:
-            self._stop_travel_if_traveling()
-
-        tilt_time = self._tilting_time_close if closing else self._tilting_time_open
-        movement_time = (abs(target - current) / 100.0) * tilt_time
-
-        travel_target = None
-        pre_step_delay = 0.0
-        needs_travel_pre_step = False
-        if self._tilt_strategy is not None:
-            current_pos = self.travel_calc.current_position()
-            if current is not None and current_pos is not None:
-                steps = self._tilt_strategy.plan_move_tilt(target, current_pos, current)
-                travel_target = extract_coupled_travel(steps)
-                pre_step_delay = calculate_pre_step_delay(
-                    steps, self._tilt_strategy, self.tilt_calc, self.travel_calc
-                )
-                if self._tilt_strategy.uses_tilt_motor and has_travel_pre_step(steps):
-                    needs_travel_pre_step = True
-
-        if self._is_movement_too_short(
-            movement_time,
-            target,
-            current,
-            "set_tilt_position",
-            is_recalibrated_leg=not recalibrate,
-        ):
-            await self._release_displaced_tilt_motor(was_tilt_motor_move)
-            return
-
-        self._require_movement_target_available(self._tilt_movement_target(command))
-        # The travel pre-step below isn't covered by the tilt gate above — see
-        # _require_movement_target_available.
-        if needs_travel_pre_step and travel_target is not None:
-            await self._start_travel_pre_step(travel_target, target, command)
-            return
-
-        self._last_command = command
-        # Set only now that the move is committed (past every no-op/too-short
-        # early return) — a tilt move must not run on at a travel endpoint, but
-        # leaking this onto a bailed-out call would wrongly suppress an
-        # in-flight travel move's run-on (issue #125).
-        self._moving_tilt = True
-
-        if self._tilt_strategy is not None and self._tilt_strategy.uses_tilt_motor:
-            self._moving_tilt_motor = True
-            if closing:
-                await self._send_tilt_close()
-            else:
-                await self._send_tilt_open()
-        else:
-            await self._async_handle_command(command)
-        self._begin_movement(
-            target,
-            travel_target,
-            self.tilt_calc,
-            self.travel_calc,
-            self._tilt_startup_delay,
-            pre_step_delay,
-        )
-
-    async def _plan_tilt_for_travel(
-        self, target: int, command: str, current_pos, current_tilt
-    ) -> tuple[int | None, float, bool]:
-        """Plan tilt coupling for a travel movement.
-
-        Returns (tilt_target, pre_step_delay, started_pre_step).
-        If started_pre_step is True, the caller should return immediately
-        because _start_tilt_pre_step has taken over the movement lifecycle.
-        """
-        tilt_target = None
-        pre_step_delay = 0.0
-        self._tilt_restore_target = None
-
-        if self._tilt_strategy is None:
-            return tilt_target, pre_step_delay, False
-
-        if current_pos is None or current_tilt is None:
-            return tilt_target, pre_step_delay, False
-
-        steps = self._tilt_strategy.plan_move_position(
-            target, current_pos, current_tilt
-        )
-        tilt_target = extract_coupled_tilt(steps)
-        pre_step_delay = calculate_pre_step_delay(
-            steps, self._tilt_strategy, self.tilt_calc, self.travel_calc
-        )
-
-        # Dual motor: tilt to safe position first, then travel
-        if (
-            tilt_target is not None
-            and self._tilt_strategy.uses_tilt_motor
-            and current_tilt != tilt_target
-        ):
-            if target in (0, 100):
-                restore = target
-                if target == 0 and not self._close_includes_tilt:
-                    # close_includes_tilt off: close travels only; the slats
-                    # stay at the safe position the pre-step drove them to.
-                    restore = tilt_target
-            elif self._tilt_strategy.allows_tilt_at_position(target):
-                restore = current_tilt
-            else:
-                restore = tilt_target  # stay at safe position
-            await self._start_tilt_pre_step(tilt_target, target, command, restore)
-            return tilt_target, pre_step_delay, True
-
-        # Dual motor: pre-step skipped, but still snap tilt to endpoint.
-        # Same close_includes_tilt guard as above: on a close with the option
-        # off, don't schedule a restore to 0 — leave tilt at the safe position.
-        if (
-            tilt_target is not None
-            and self._tilt_strategy.uses_tilt_motor
-            and target in (0, 100)
-            and current_tilt != target
-            and not (target == 0 and not self._close_includes_tilt)
-        ):
-            self._tilt_restore_target = target
-
-        # Shared motor with restore: save tilt for post-travel restore
-        if (
-            tilt_target is not None
-            and self._tilt_strategy.restores_tilt
-            and not self._tilt_strategy.uses_tilt_motor
-            and target not in (0, 100)
-        ):
-            self._tilt_restore_target = current_tilt
-
-        return tilt_target, pre_step_delay, False
-
-    async def _handle_pre_movement_checks(self, command):
-        """Handle startup delay conflicts and relay delay before a movement.
-
-        Returns (should_proceed, is_direction_change).
-        """
-        is_direction_change = self._is_direction_change(command)
-
-        # If startup delay active for same direction, don't restart
-        if self._startup_delay_task and not self._startup_delay_task.done():
-            if not is_direction_change:
-                self._log(
-                    "_handle_pre_movement_checks :: startup delay active, skipping"
-                )
-                return False, is_direction_change
-            # On tap hardware the stop waits the relay confirmation out, and
-            # the confirmation runs the parked start: the move is then in
-            # motion, and the caller's own in-motion direction-change block
-            # stops it, settles, and drives the other way. Only a start that is
-            # still parked afterwards is cancelled unrun and stopped here.
-            await self._await_confirmation_before_stop()
-            if self._startup_delay_task and not self._startup_delay_task.done():
-                self._log(
-                    "_handle_pre_movement_checks :: direction change, cancelling startup delay"
-                )
-                self._cancel_startup_delay_task()
-                await self._async_handle_command(SERVICE_STOP_COVER)
-            else:
-                self._log(
-                    "_handle_pre_movement_checks :: direction change, the confirmed"
-                    " move is reversed in motion"
-                )
-
-        return True, is_direction_change
-
-    def _is_movement_too_short(
-        self,
-        movement_time,
-        target,
-        current,
-        label,
-        *,
-        is_recalibrated_leg: bool = False,
-    ):
-        """Check if movement time is below minimum. Returns True if movement should be skipped.
-
-        ``is_recalibrated_leg`` marks leg B of a recalibrated move (#179) — the
-        caller passed ``recalibrate=False``. min_movement_time exists to skip
-        pointless motor pulses for imperceptible moves, but by the time leg B
-        runs, leg A has already driven the motor a full travel to reach a true
-        datum; the second pulse to the position the user actually asked for IS
-        the point of the operation, not a redundant nudge. Rejecting it here
-        would silently strand the cover at leg A's endpoint instead of where
-        it was asked to go, so leg B is exempt from this check entirely.
-        """
-        if is_recalibrated_leg:
-            return False
-        is_to_endpoint = target in (0, 100)
-        if (
-            self._min_movement_time is not None
-            and self._min_movement_time > 0
-            and not is_to_endpoint
-            and movement_time < self._min_movement_time
-        ):
-            _LOGGER.info(
-                "%s :: movement too short (%fs < %fs), ignoring - from %d%% to %d%%",
-                label,
-                movement_time,
-                self._min_movement_time,
-                current,
-                target,
-            )
-            self.async_write_ha_state()
-            return True
-        return False
-
-    def _require_configured(self) -> None:
-        """Raise if the cover is not properly configured."""
-        missing = self._get_missing_configuration()
-        if missing:
-            raise HomeAssistantError(
-                f"Cover not configured: missing {', '.join(missing)}. "
-                "Please configure using the Cover Time Based card."
-            )
-
-    def _require_travel_time(self, closing: bool) -> float:
-        """Return travel time for the given direction, or raise if not configured."""
-        travel_time = self._travel_time_close if closing else self._travel_time_open
-        if travel_time is None:
-            raise HomeAssistantError(
-                "Travel time not configured. Please configure travel times "
-                "using the Cover Time Based card."
-            )
-        return travel_time
-
-    def _are_entities_configured(self) -> bool:
-        """Return True if the required input entities are configured.
-
-        Subclasses override this to check their specific entity IDs.
-        """
-        return True
-
-    def _get_missing_configuration(self) -> list[str]:
-        """Return list of missing configuration items."""
-        missing = []
-        if not self._are_entities_configured():
-            missing.append(self._missing_entities_label)
-        if self._travel_time_close is None and self._travel_time_open is None:
-            missing.append("travel times")
-        return missing
-
-    def _target_entity_ids(self) -> list[str]:
-        """Return the configured target entity IDs this cover drives.
-
-        These are the switch/button/script entities that actually move the
-        device. Subclasses (e.g. wrapped mode) extend this with their own
-        targets.
-        """
-        return [
-            entity_id
-            for attr in self._SWITCH_TARGET_ATTRS
-            if (entity_id := getattr(self, attr, None))
-        ]
-
-    @staticmethod
-    def _entity_unavailable(state) -> bool:
-        """Return True if a target entity is unavailable.
-
-        Unavailable means the entity is missing (state is None) or its state is
-        STATE_UNAVAILABLE. STATE_UNKNOWN is treated as available (a connected
-        entity whose value isn't known yet, e.g. a button after a restart).
-        """
-        return state is None or state.state == STATE_UNAVAILABLE
-
-    def _any_target_unavailable(self) -> bool:
-        """Short-circuiting check used by the hot-path `available` property."""
-        if self.hass is None:
-            return False
-        return any(
-            self._entity_unavailable(self.hass.states.get(entity_id))
-            for entity_id in self._target_entity_ids()
-        )
-
-    def _movement_target(self, closing: bool) -> str | None:
-        """Entity driven to move the cover in the given travel direction."""
-        return self._close_switch_entity_id if closing else self._open_switch_entity_id
-
-    def _tilt_movement_target(self, command: str) -> str | None:
-        """Entity driven for a resolved tilt command.
-
-        ``command`` is what the funnel already computed via ``tilt_command_for``
-        (which some strategies, e.g. sequential_open, invert). Dual-motor tilt
-        uses dedicated tilt switches; shared-motor tilt drives the main travel
-        switch.
-        """
-        closing = command == SERVICE_CLOSE_COVER
-        if self._tilt_strategy is not None and self._tilt_strategy.uses_tilt_motor:
-            return self._tilt_close_switch_id if closing else self._tilt_open_switch_id
-        return self._movement_target(closing)
-
-    def _require_movement_target_available(self, target: str | None) -> None:
-        """Reject a fresh, self-initiated movement whose target is unavailable.
-
-        Only fresh user/automation movement is gated (`_self_initiated_movement`).
-        Stops, reverses, external reactions, retargets, and internal
-        continuations never reach here (or have `_self_initiated_movement` False),
-        so the cover can always be halted regardless of target availability.
-
-        Movement that runs via a pre-step (dual-motor / sequential) returns from
-        the funnel before reaching the funnel's own per-direction gate, so each
-        pre-step phase calls this directly for its target (in
-        `_start_tilt_pre_step` / `_start_travel_pre_step`). The all-targets
-        `available` flag (the cover reports unavailable whenever any target is
-        down) is the backstop covering both paths.
-        """
-        if not self._self_initiated_movement or self.hass is None:
-            return
-        if target and self._entity_unavailable(self.hass.states.get(target)):
-            raise HomeAssistantError(
-                f"Cover target '{target}' is unavailable; cannot start movement."
-            )
-
-    def _has_tilt_support(self):
-        """Return if cover has tilt support."""
-        return (
-            self.supports_tilt
-            and self._tilt_strategy is not None
-            and hasattr(self, "tilt_calc")
-        )
-
-    # -----------------------------------------------------------------------
     # Movement tracking
     # -----------------------------------------------------------------------
 
@@ -3082,929 +2029,6 @@ class CoverTimeBased(CalibrationMixin, SwitchEchoMixin, CoverEntity, RestoreEnti
         return self.travel_calc.position_reached(travel_pos) and (
             not self._has_tilt_support() or self.tilt_calc.position_reached(tilt_pos)
         )
-
-    # -----------------------------------------------------------------------
-    # Movement lifecycle (auto-stop, pre-step, restore)
-    # -----------------------------------------------------------------------
-
-    async def auto_stop_if_necessary(self):
-        """Do auto stop if necessary.
-
-        ``auto_updater_hook`` spawns this only on arrival, so anything added
-        outside the ``position_reached`` check below would never run for it.
-        """
-        if self.position_reached():
-            self._log(
-                "auto_stop_if_necessary :: position reached (self_initiated=%s)",
-                self._self_initiated_movement,
-            )
-            self.travel_calc.stop()
-            if self._has_tilt_support():
-                self.tilt_calc.stop()
-
-            if not self._self_initiated_movement:
-                # Movement was triggered externally. A multi-phase move still
-                # has to chain its phases (dual-motor tilt pre-step → travel, or
-                # travel pre-step → tilt); only the final endpoint relay handling
-                # is special-cased below. Without continuing here the cover would
-                # tilt-to-safe and then stall, needing a second press to travel.
-                self._log("auto_stop_if_necessary :: external movement")
-                if self._pending_travel_target is not None:
-                    self._log(
-                        "auto_stop_if_necessary :: external tilt pre-step complete"
-                    )
-                    await self._start_pending_travel()
-                    return
-                if self._pending_tilt_target is not None:
-                    self._log(
-                        "auto_stop_if_necessary :: external travel pre-step complete"
-                    )
-                    await self._start_pending_tilt()
-                    return
-                # Move complete: don't re-drive the relay, but a latched relay
-                # (switch mode) must still be de-energized; momentary modes
-                # self-released and no-op in _settle_external_endpoint.
-                was_tilt_motor_move = self._moving_tilt_motor
-                await self._settle_external_endpoint()
-                if self._tilt_strategy is not None:
-                    self._tilt_strategy.snap_trackers_to_physical(
-                        self.travel_calc, self.tilt_calc
-                    )
-                self._last_command = None
-                self._moving_tilt_motor = False
-                self._moving_tilt = False
-                if was_tilt_motor_move:
-                    self._on_tilt_motor_move_complete()
-                # The one terminal return here that does NOT dispatch an armed
-                # recalibration leg — drop it instead (issue #179). Dispatching
-                # would be wrong: leg B is a self-initiated move chained off a
-                # datum leg A reached, and this branch is the *hardware* having
-                # finished a move of its own. Driving the motor again off the
-                # back of that is exactly the surprise-move the feature promises
-                # never to make. A leg should not be armed here at all — an
-                # external trigger gets RecalibrationPlan.NONE, and every route
-                # that flips _self_initiated_movement to False runs
-                # _abandon_active_lifecycle first, which clears the pending — so
-                # this makes that invariant local and enforced rather than
-                # inferred from four call sites. Self-limiting either way; the
-                # point is that it cannot rot.
-                self._disarm_recalibrated_leg()
-                await self._async_persist_position()
-                return
-
-            if self._tilt_restore_active:
-                self._log("auto_stop_if_necessary :: tilt restore complete")
-                self._release_tilt_restore()
-                if self._has_tilt_motor():
-                    await self._tilt_settle()
-                    self._on_tilt_motor_move_complete()
-                else:
-                    await self._async_handle_command(SERVICE_STOP_COVER)
-                if self._tilt_strategy is not None:
-                    self._tilt_strategy.snap_trackers_to_physical(
-                        self.travel_calc, self.tilt_calc
-                    )
-                self._last_command = None
-                self._moving_tilt_motor = False
-                self._moving_tilt = False
-                await self._async_persist_position()
-                await self._maybe_start_recalibrated_leg()
-                return
-
-            if self._pending_travel_target is not None:
-                # Tilt pre-step complete — start travel phase
-                self._log("auto_stop_if_necessary :: tilt pre-step complete")
-                await self._start_pending_travel()
-                return
-
-            if self._pending_tilt_target is not None:
-                # Travel pre-step complete — start tilt phase
-                self._log("auto_stop_if_necessary :: travel pre-step complete")
-                await self._start_pending_tilt()
-                return
-
-            if self._tilt_strategy is not None:
-                self._tilt_strategy.snap_trackers_to_physical(
-                    self.travel_calc, self.tilt_calc
-                )
-
-            if self._tilt_restore_target is not None:
-                # Travel just completed — start tilt restore phase. If the
-                # restore target already matches where tilt is (issue #179:
-                # leg A always targets 100, which coincides with the default
-                # safe_tilt_position, so the dual-motor pre-step already
-                # parked tilt exactly at the restore target), _start_tilt_restore
-                # returns synchronously without driving a motor or re-arming
-                # the auto-updater — this IS the terminal completion of the
-                # move and must consume any armed recalibration leg. When a
-                # real restore is claimed instead (_tilt_restore_active
-                # becomes True), its own eventual completion is handled by
-                # the _tilt_restore_active branch above — checking here too
-                # would double-fire.
-                await self._start_tilt_restore()
-                if not self._tilt_restore_active:
-                    # Mirrors the other two leg-B dispatch sites (issue #179):
-                    # persist before consuming any armed recalibration leg,
-                    # for consistency across the three otherwise-parallel
-                    # sites. Impact is nil either way — leg B persists on its
-                    # own completion — but the asymmetry reads as a bug later
-                    # if left alone.
-                    await self._async_persist_position()
-                    await self._maybe_start_recalibrated_leg()
-                return
-
-            current_travel = self.travel_calc.current_position()
-            # Endpoint handling (the self-stop skip and the run-on below) is a
-            # *travel* concept: it only applies when a travel move reaches a
-            # physical limit. A tilt move that merely finishes while the cover is
-            # parked at a travel endpoint has driven the motor *off* that limit
-            # to articulate the slats, so the motor will NOT self-stop there and
-            # must always be stopped explicitly — never given the self-stop skip
-            # (issue #142) or the run-on (issue #125).
-            endpoint_applies = (
-                self._at_endpoint(current_travel) and not self._moving_tilt
-            )
-            if self._motor_stops_itself():
-                # The device drives to the target and holds there on its own
-                # (e.g. a wrapped cover commanded via set_cover_position). Any
-                # stop we issue here is at best redundant and at worst nudges it
-                # off the exact target (a freeze re-commands the calculated, not
-                # requested, position). Just settle the tracker.
-                self._log("auto_stop_if_necessary :: device self-stops, no relay stop")
-            elif self._has_tilt_motor() and self._moving_tilt_motor:
-                # The completed movement drove the dedicated tilt motor — settle
-                # that motor (skipping the stop at the tilt endpoints), not
-                # travel. Without this a tilt move would fall through to the
-                # travel stop below and re-pulse the travel relay off a stale
-                # _last_command.
-                await self._tilt_settle()
-                self._on_tilt_motor_move_complete()
-            elif endpoint_applies and self._self_stops_at_endpoints():
-                # The motor self-stops at its physical limit switch. Sending a
-                # stop here is redundant (and for toggle re-pulses → restart),
-                # so skip the relay stop and any run-on; just settle the
-                # tracker. The run-on (below) is therefore reached only by
-                # hardware that does NOT self-stop at its endpoints — switch
-                # mode, and pulse with send_endpoint_stop enabled.
-                self._log(
-                    "auto_stop_if_necessary :: motor self-stops at endpoint"
-                    " (position=%d), no relay stop",
-                    current_travel,
-                )
-            elif (
-                endpoint_applies
-                and self._endpoint_runon_time is not None
-                and self._endpoint_runon_time > 0
-            ):
-                self._log(
-                    "auto_stop_if_necessary :: at endpoint (position=%d),"
-                    " delaying relay stop by %fs",
-                    current_travel,
-                    self._endpoint_runon_time,
-                )
-                self._delay_task = self.hass.async_create_task(
-                    self._delayed_stop(self._endpoint_runon_time)
-                )
-            else:
-                await self._async_handle_command(SERVICE_STOP_COVER)
-            if endpoint_applies and not self._moving_tilt_motor:
-                # endpoint_applies implies _at_endpoint(current_travel), which is
-                # False for None, so the position is known here.
-                assert current_travel is not None
-                self._on_endpoint_reached(int(current_travel))
-            self._last_command = None
-            self._moving_tilt_motor = False
-            self._moving_tilt = False
-            await self._async_persist_position()
-            await self._maybe_start_recalibrated_leg()
-
-    def _motor_stops_itself(self) -> bool:
-        """Return True if the device halts at the target without a stop command.
-
-        Relay-driven covers (the default) must be told to stop when the tracker
-        reaches the target, so this is False. Subclasses whose device has native
-        position control and stops itself at the commanded position override this
-        to True so auto-stop skips the redundant (and potentially target-nudging)
-        relay stop.
-        """
-        return False
-
-    def _self_stops_at_endpoints(self) -> bool:
-        """Return True if the motor self-stops at the physical endpoints.
-
-        Roller-shutter motors have internal limit switches that halt the motor
-        at fully-open/closed on their own. For modes whose relays are momentary
-        or delegated (toggle, pulse, wrapped), the stop we would otherwise send
-        at 0%/100% is therefore redundant — and for toggle actively harmful, as
-        re-pulsing the direction relay restarts the already-stopped motor. Such
-        modes return True so auto-stop skips the relay stop (and run-on) at an
-        endpoint while still stopping mid-travel, where nothing self-stops.
-
-        Switch mode overrides this to False: its direction relay is latched ON
-        for the whole travel, so reaching an endpoint must still de-energize it.
-        """
-        return True
-
-    def _supports_stepped_calibration(self) -> bool:
-        """Whether the overhead and minimum-movement tests can run on this mode.
-
-        Both restart the motor in the same direction from a stop, which every
-        relay-driven mode can do. A mode whose restart-after-stop is a
-        reversal overrides this to False.
-        """
-        return True
-
-    def _at_endpoint(self, position) -> bool:
-        """Return True at a travel endpoint (0/100) where endpoint handling
-        (self-stop skip or run-on) applies.
-
-        Sequential tilt drives the motor past cover-closed to articulate the
-        slats, so it disallows endpoint handling at 0 (``allows_endpoint_runon``).
-        """
-        return (
-            position is not None
-            and position in (0, 100)
-            and (
-                self._tilt_strategy is None
-                or self._tilt_strategy.allows_endpoint_runon(position)
-            )
-        )
-
-    async def _tilt_settle(self) -> None:
-        """Stop the dedicated tilt motor at the end of a tilt-motor movement.
-
-        Mirrors the travel endpoint logic: at the tilt endpoints (0%/100%) the
-        tilt motor self-stops on its own limit, so the stop is skipped (and for
-        toggle a re-pulse would restart it) — except in switch mode, whose
-        latched tilt relay must still be de-energized. Mid-tilt always stops.
-        """
-        current_tilt = (
-            self.tilt_calc.current_position() if self._has_tilt_support() else None
-        )
-        if (
-            current_tilt is not None
-            and current_tilt in (0, 100)
-            and self._self_stops_at_endpoints()
-        ):
-            self._log(
-                "_tilt_settle :: tilt motor self-stops at endpoint (%d), no relay stop",
-                current_tilt,
-            )
-        else:
-            await self._send_tilt_stop()
-
-    def _on_tilt_motor_move_complete(self) -> None:
-        """Hook: a dedicated-tilt-motor movement finished (settled or skipped).
-
-        Called from ``auto_stop_if_necessary`` after a movement that drove the
-        dedicated tilt motor completes, whether self-initiated (right after
-        ``_tilt_settle``) or externally triggered. The base class has no tilt
-        bookkeeping of its own to clear; toggle-style hardware overrides this
-        to clear ``_last_tilt_direction`` so a stale direction can't outlive
-        the move that set it (see ``_abandon_active_lifecycle``, which pulses
-        a relay keyed off that direction).
-        """
-
-    async def _settle_external_endpoint(self) -> None:
-        """De-energize any latched relay after an externally-triggered move
-        reaches its endpoint.
-
-        Auto-stop skips the relay stop for externally-triggered movements
-        (``_self_initiated_movement`` False): the trigger came from outside HA,
-        and for momentary modes (pulse/toggle/wrapped) the relay was a brief
-        pulse that has already self-released — there is nothing to de-energize,
-        so they keep this no-op. Switch mode latches its direction relay ON for
-        the whole travel, so it overrides this to turn the relay off (only if
-        still on); otherwise the relay stays energized at the endpoint forever.
-        """
-        return
-
-    def _park_axis_at_limit(self, calc: TravelCalculator, limit: int) -> None:
-        """Record an axis at the limit its motor will reach on its own.
-
-        Used when this entity stops tracking a motor it cannot stop (removal
-        on hardware that self-stops at its limits). Modes that carry more
-        state than the tracker anchor it here too.
-        """
-        calc.set_position(limit)
-
-    def _settle_axis_after_removal(
-        self,
-        calc: TravelCalculator,
-        *,
-        driving: bool,
-        fallback_limit: int | None,
-    ) -> None:
-        """Leave one axis's tracker where the motor will actually be.
-
-        Removal stops every motor it can stop, so an axis is still running
-        afterwards only on hardware that halts at its own limits. Never
-        driving, stopped by removal, or a device that holds its own target:
-        the tracker's own position stands. Otherwise park at the limit in the
-        direction of travel — from the tracker if it is travelling, else from
-        the last command — or forget the axis when no direction is known.
-        """
-        if (
-            not driving
-            or not self._self_stops_at_endpoints()
-            or self._motor_stops_itself()
-        ):
-            calc.stop()
-            return
-        if calc.is_traveling():
-            limit = 100 if calc.travel_direction == TravelStatus.DIRECTION_UP else 0
-            self._park_axis_at_limit(calc, limit)
-        elif fallback_limit is not None:
-            self._park_axis_at_limit(calc, fallback_limit)
-        else:
-            calc.clear_position()
-
-    def _limit_for_last_command(self) -> int | None:
-        if self._last_command == SERVICE_OPEN_COVER:
-            return 100
-        if self._last_command == SERVICE_CLOSE_COVER:
-            return 0
-        return None
-
-    def _tilt_limit_for_last_command(self) -> int | None:
-        """The tilt limit the command now running heads for, or None.
-
-        The tilt tracker is idle across a deferred start, so the direction can
-        only come from the command already at the relay. Inverting
-        ``tilt_command_for`` — the same mapping the move used to choose that
-        command — keeps strategies whose slats open by driving the motor the
-        other way (sequential_open) pointing at the tilt limit rather than the
-        motor's.
-        """
-        if self._last_command is None:
-            return None
-        strategy = self._tilt_strategy
-        if strategy is None:
-            return self._limit_for_last_command()
-        if self._last_command == strategy.tilt_command_for(True):
-            return 0
-        if self._last_command == strategy.tilt_command_for(False):
-            return 100
-        return None
-
-    def _on_endpoint_reached(self, endpoint: int) -> None:
-        """Hook: a self-initiated travel move terminated at 0 or 100.
-
-        Base is a no-op. Overridden by modes that must re-anchor internal
-        state at a physical limit (e.g. single-button phase tracking).
-        """
-        return
-
-    async def _delayed_stop(self, delay):
-        """Stop the relay after a delay.
-
-        Clears ``_delay_task`` only while it still points at this task: a
-        cancel's cleanup runs a loop turn later, by which time a replacement
-        run-on may already be registered, and clearing that would hide a
-        pending relay stop from _cancel_delay_task and _movement_in_progress.
-        """
-        self._log("_delayed_stop :: waiting %fs before stopping relay", delay)
-        try:
-            await sleep(delay)
-            self._log("_delayed_stop :: delay complete, stopping relay")
-            await self._async_handle_command(SERVICE_STOP_COVER)
-            self._last_command = None
-        except asyncio.CancelledError:
-            self._log("_delayed_stop :: delay cancelled")
-            raise
-        finally:
-            if self._delay_task is asyncio.current_task():
-                self._delay_task = None
-
-    async def _stop_travel_relay_if_needed(self, *, travel_was_running: bool) -> None:
-        """Send the travel STOP unless it would be a movement command.
-
-        Momentary hardware self-stops at its limits, and a 'stop' pulse to a
-        stopped motor is a movement command (#153) or a go-to-favourite (#133).
-        Skip when the travel axis never ran under this lifecycle, or when it is
-        settled at an endpoint the motor self-stopped at. Switch mode
-        (_self_stops_at_endpoints False) always de-energizes.
-        """
-        if not self._self_stops_at_endpoints():
-            await self._async_handle_command(SERVICE_STOP_COVER)
-            return
-        if not travel_was_running:
-            self._log("_stop_travel_relay_if_needed :: travel idle, skipping stop")
-            return
-        if self._at_endpoint(self.travel_calc.current_position()):
-            self._log(
-                "_stop_travel_relay_if_needed :: self-stopped at endpoint,"
-                " skipping stop"
-            )
-            return
-        await self._async_handle_command(SERVICE_STOP_COVER)
-
-    async def _abandon_active_lifecycle(self):
-        """Abandon any active multi-phase tilt lifecycle (pre-step, restore).
-
-        Called at the start of every movement method. If a tilt restore or
-        tilt pre-step is in progress, stops all hardware and calculators.
-        Always clears the pending restore target so it won't fire after
-        the next travel completes.
-
-        Being the one hook every movement entry point funnels through, this is
-        also where a new movement claims the epoch — see
-        _settle_before_reversing. Tilt counts: on shared-motor strategies tilt
-        is driven by the travel motor, so a tilt command must invalidate a
-        travel reversal parked in its settle gap just as a travel command does.
-        """
-        self._supersede_movement()
-        # A fresh move must not inherit a relay-feedback arm left set by an
-        # earlier operation that armed but never reached _begin_movement (e.g.
-        # an endpoint resync). Otherwise a following suppress_start_command
-        # redrive — which fires no _send_* to re-arm — would read the stale arm
-        # and wait for an ON echo its already-latched relay will never send.
-        self._feedback_armed_entity = None
-        was_restoring = self._tilt_restore_active
-        was_pre_stepping = (
-            self._pending_travel_target is not None
-            or self._pending_tilt_target is not None
-        )
-        # Captured before the resets below clear them: the tilt stop at the
-        # bottom of this method must reflect whether the tilt motor was
-        # actually driven, not whatever _moving_tilt_motor/tilt_calc read
-        # after being reset here — an idle-motor pulse off a stale
-        # _last_tilt_direction is a #153-class hazard (audit finding B4).
-        was_tilt_motor = self._moving_tilt_motor
-        was_tilt_traveling = self._has_tilt_support() and self.tilt_calc.is_traveling()
-        # Captured before the calculator stops below: the travel stop at the
-        # bottom must reflect whether the travel motor was actually driven under
-        # this lifecycle. A dual-motor tilt pre-step never starts the travel
-        # motor, so its abandon must not pulse a travel relay off a stale
-        # _last_command at an idle motor (#153 / #133, audit Task 5B). A pending
-        # run-on delay counts as still-driving (any mode with an endpoint run-on
-        # arms _delay_task — e.g. pulse — not switch mode only).
-        travel_was_running = self.travel_calc.is_traveling() or (
-            self._delay_task is not None and not self._delay_task.done()
-        )
-
-        # Always clear multi-phase state
-        self._clear_multiphase_tilt_state()
-        # Each movement entry point funnels through here; default to a travel
-        # move and let the tilt paths below opt in.
-        self._moving_tilt_motor = False
-        self._moving_tilt = False
-
-        if not was_restoring and not was_pre_stepping:
-            return
-
-        self._log(
-            "_abandon_active_lifecycle :: abandoning %s",
-            "tilt restore" if was_restoring else "pre-step",
-        )
-
-        self._cancel_startup_delay_task()
-
-        if self.travel_calc.is_traveling():
-            self.travel_calc.stop()
-        if self._has_tilt_support() and self.tilt_calc.is_traveling():
-            self.tilt_calc.stop()
-        self.stop_auto_updater()
-
-        await self._stop_travel_relay_if_needed(travel_was_running=travel_was_running)
-        # Deliberately still gated on _triggered_externally, unlike
-        # async_stop_cover. This route cannot tell which axis reported: the
-        # tilt entry points reach it too (async_open_cover_tilt ->
-        # _async_move_tilt_to_endpoint -> here), so firing on an external tilt
-        # event would echo a stop back at the very relay that reported — which
-        # on toggle hardware starts the motor rather than stopping it. Until
-        # the axis is threaded this far, suppressing is the safe direction.
-        if (
-            self._has_tilt_motor()
-            and not self._triggered_externally
-            and (was_tilt_motor or was_tilt_traveling)
-        ):
-            await self._send_tilt_stop()
-
-    def _stop_travel_if_traveling(self):
-        """Stop cover movement if it's currently traveling."""
-        if self.travel_calc.is_traveling():
-            self._log("_stop_travel_if_traveling :: stopping cover movement")
-            self.travel_calc.stop()
-            if self._has_tilt_support() and self.tilt_calc.is_traveling():
-                self._log("_stop_travel_if_traveling :: also stopping tilt")
-                self.tilt_calc.stop()
-
-    def _handle_stop(self, *, supersede: bool = True):
-        """Handle stop.
-
-        ``supersede`` says whether this is a fresh command to halt the cover,
-        which claims the movement and so keeps a reversal parked in its settle
-        gap from driving afterwards. Every route that halts the cover lands
-        here — through _neutralize_tracked_movement for both _stop_hardware
-        implementations (CoverTimeBased's and ToggleBaseCover's, which does not
-        call super), calibration start, a raw command and a direction change
-        at a parked start (_neutralize_parked_move, which stops and returns
-        rather than driving on, so it supersedes like a stop), and directly for
-        the known-position declarations.
-
-        Passive routes must pass ``supersede=False``: a wrapped cover reporting
-        its settled position snaps via set_known_position, and a switch-mode
-        relay's unmarked off (a hardware interlock clearing the opposite relay)
-        calls async_stop_cover. Both can arrive inside the settle window — the
-        wrapped self-echo suppressors are keyed on is_traveling(), which the
-        reversal has already cleared — and treating them as supersessions would
-        silently drop the user's move, or freeze the tracker while the motor
-        runs. So must the stop a reversal issues as its own prelude, which would
-        otherwise cancel the very movement it is starting.
-
-        The caller has to say so because _triggered_externally cannot: it is
-        ambient instance state held for the whole external call, settle gap
-        included, so inferring from it read a genuine stop landing in that gap
-        as a device echo and dropped it.
-        """
-        if supersede:
-            self._supersede_movement()
-        self._clear_multiphase_tilt_state()
-
-        if self.travel_calc.is_traveling():
-            self._log("_handle_stop :: button stops cover movement")
-            self.travel_calc.stop()
-            self.stop_auto_updater()
-
-        if self._has_tilt_support() and self.tilt_calc.is_traveling():
-            self._log("_handle_stop :: button stops tilt movement")
-            self.tilt_calc.stop()
-            self.stop_auto_updater()
-
-    def _neutralize_tracked_movement(self, *, supersede: bool = True) -> None:
-        """Park the tracker and cancel deferred starts, without touching a relay.
-
-        The prelude every route that takes the movement away from the tracker
-        shares: a deferred start left armed would later drive the motor, or fire
-        a relay stop, against whatever took over.
-
-        See _handle_stop for ``supersede``.
-        """
-        self._cancel_startup_delay_task()
-        self._cancel_delay_task()
-        self._handle_stop(supersede=supersede)
-
-    async def _start_tilt_pre_step(
-        self, tilt_target, travel_target, travel_command, restore_target
-    ):
-        """Move tilt to safe position before travel (dual_motor).
-
-        Sends the tilt motor command and starts tilt_calc. When tilt reaches
-        target, auto_stop_if_necessary will call _start_pending_travel to
-        begin the actual cover travel.
-        """
-        current_tilt = self.tilt_calc.current_position()
-        self._log(
-            "_start_tilt_pre_step :: tilt %s→%d, pending travel→%d (%s)",
-            current_tilt,
-            tilt_target,
-            travel_target,
-            travel_command,
-        )
-        closing_tilt = tilt_target < current_tilt
-        self._require_movement_target_available(
-            self._tilt_movement_target(
-                SERVICE_CLOSE_COVER if closing_tilt else SERVICE_OPEN_COVER
-            )
-        )
-        self._pending_travel_target = travel_target
-        self._pending_travel_command = travel_command
-        self._tilt_restore_target = restore_target
-
-        if not self._triggered_externally:
-            if closing_tilt:
-                await self._send_tilt_close()
-            else:
-                await self._send_tilt_open()
-
-        if self._removed:
-            return
-        self.tilt_calc.start_travel(tilt_target)
-        self.start_auto_updater()
-
-    async def _start_pending_travel(self):
-        """Start travel after tilt pre-step completes (dual_motor).
-
-        Called by auto_stop_if_necessary when tilt_calc reaches the safe
-        position. For self-initiated moves: stops the tilt motor and sends
-        the travel command. For external triggers (where hardware is doing
-        the multi-phase motion itself), only updates the integration's
-        trackers — the relays are left to the hardware.
-
-        The origin gate is the instance flag ``_self_initiated_movement``,
-        NOT ``_triggered_externally``: this continuation runs on the
-        auto-updater's own task (auto_updater_hook schedules it via
-        hass.async_create_task), and ``_triggered_externally`` is task-scoped,
-        so it is always False here regardless of how the move began.
-        ``_self_initiated_movement`` is set once at the entry point and
-        survives across the phase boundary, so it is the correct carrier.
-        """
-        target = self._pending_travel_target
-        command = self._pending_travel_command
-        assert target is not None and command is not None
-        self._pending_travel_target = None
-        self._pending_travel_command = None
-
-        self._log(
-            "_start_pending_travel :: starting travel to %d (%s)",
-            target,
-            command,
-        )
-
-        self._moving_tilt_motor = False
-        if self._self_initiated_movement:
-            # Stop tilt motor and send travel command.
-            await self._send_tilt_stop()
-            await self._async_handle_command(command)
-        self._last_command = command
-        self._begin_movement(
-            target,
-            None,
-            self.travel_calc,
-            None,
-            self._travel_startup_delay,
-        )
-
-    async def _start_travel_pre_step(self, travel_target, tilt_target, tilt_command):
-        """Move cover to allowed position before tilt (dual_motor).
-
-        Sends the travel motor command and starts travel_calc. When travel
-        reaches target, auto_stop_if_necessary will call _start_pending_tilt
-        to begin the actual tilt movement.
-        """
-        current_pos = self.travel_calc.current_position()
-        self._log(
-            "_start_travel_pre_step :: travel %s→%d, pending tilt→%d (%s)",
-            current_pos,
-            travel_target,
-            tilt_target,
-            tilt_command,
-        )
-        closing = travel_target < current_pos
-        command = SERVICE_CLOSE_COVER if closing else SERVICE_OPEN_COVER
-        self._require_movement_target_available(self._movement_target(closing))
-        self._pending_tilt_target = tilt_target
-        self._pending_tilt_command = tilt_command
-        self._last_command = command
-        await self._async_handle_command(command)
-
-        self._begin_movement(
-            travel_target,
-            None,
-            self.travel_calc,
-            None,
-            self._travel_startup_delay,
-        )
-
-    async def _start_pending_tilt(self):
-        """Start tilt after travel pre-step completes (dual_motor).
-
-        Called by auto_stop_if_necessary when travel_calc reaches the
-        allowed position. For self-initiated moves: stops the travel motor,
-        sends the tilt command, and starts tracking with tilt_calc. For
-        external triggers (where the hardware is driving the multi-phase
-        motion itself), only starts tilt tracking — the relays are left to
-        the hardware.
-
-        The origin gate is the instance flag ``_self_initiated_movement``,
-        NOT ``_triggered_externally``: this continuation runs on the
-        auto-updater's own task, and ``_triggered_externally`` is task-scoped,
-        so it is always False here regardless of how the move began.
-        ``_self_initiated_movement`` is set once at the entry point and
-        survives across the phase boundary, so it is the correct carrier.
-        """
-        target = self._pending_tilt_target
-        command = self._pending_tilt_command
-        assert target is not None and command is not None
-        self._pending_tilt_target = None
-        self._pending_tilt_command = None
-
-        self._log(
-            "_start_pending_tilt :: starting tilt to %d (%s)",
-            target,
-            command,
-        )
-
-        # Stop travel motor
-        if self._self_initiated_movement:
-            await self._async_handle_command(SERVICE_STOP_COVER)
-
-        # Send tilt command and start tracking
-        self._moving_tilt_motor = True
-        closing_tilt = command == SERVICE_CLOSE_COVER
-        if self._self_initiated_movement:
-            if closing_tilt:
-                await self._send_tilt_close()
-            else:
-                await self._send_tilt_open()
-        self._last_command = command
-        self._begin_movement(
-            target,
-            None,
-            self.tilt_calc,
-            None,
-            self._tilt_startup_delay,
-        )
-
-    async def _start_tilt_restore(self):
-        """Restore tilt to its pre-movement position.
-
-        For dual_motor: stops travel motor, starts tilt motor.
-        For shared motor (inline): reverses main motor direction.
-        """
-        restore_target = self._tilt_restore_target
-        self._tilt_restore_target = None
-        if restore_target is None:
-            return
-
-        current_tilt = self.tilt_calc.current_position()
-        if current_tilt is None or current_tilt == restore_target:
-            self._log(
-                "_start_tilt_restore :: no restore needed (current=%s, target=%s)",
-                current_tilt,
-                restore_target,
-            )
-            # Restore only runs after a travel phase, so the travel axis did run;
-            # the endpoint check inside the helper skips the stop when the motor
-            # self-stopped at its limit (a stop pulse there is a #153/#133 move).
-            await self._stop_travel_relay_if_needed(travel_was_running=True)
-            self._last_command = None
-            return
-
-        self._log(
-            "_start_tilt_restore :: restoring tilt from %d%% to %d%%",
-            current_tilt,
-            restore_target,
-        )
-
-        closing = restore_target < current_tilt
-
-        # Everything below runs on the auto-updater's background task, so every
-        # await yields the event loop and a user STOP or a new movement command
-        # can interleave. Mark the restore active *before* the first await so
-        # those paths (async_stop_cover -> _handle_stop, other moves ->
-        # _abandon_active_lifecycle) recognise it, stop the motor, and clear the
-        # flag; we then re-check the flag after each await and bail so we never
-        # (re)start or keep tracking a motor the user just stopped. The
-        # auto-updater stays unsubscribed for this whole window (stopped when
-        # travel reached its target, re-armed only at the tail below), so no
-        # re-entrant auto_stop_if_necessary can take the restore-complete branch
-        # mid-startup — keep it that way if this ever moves.
-        epoch = self._claim_tilt_restore()
-
-        if self._tilt_strategy.uses_tilt_motor:
-            # Dual motor: stop travel, then start the separate tilt motor. The
-            # travel motor self-stopped if travel just reached an endpoint, so
-            # gate the stop (a pulse there re-opens the cover untracked — #153).
-            await self._stop_travel_relay_if_needed(travel_was_running=True)
-            # Clear only after the stop helper uses _last_command to pick its
-            # relay. Tilt now owns the motion; a stale travel command would
-            # make a mid-restore tilt command pulse the parked travel motor.
-            self._last_command = None
-            if self._tilt_restore_superseded(epoch):
-                self._log("_start_tilt_restore :: cancelled before tilt motor start")
-                return
-            if closing:
-                await self._send_tilt_close()
-            else:
-                await self._send_tilt_open()
-        else:
-            # Shared motor (inline tilt — the only strategy that both restores
-            # tilt and drives it via the travel motor): the travel motor is
-            # still running from the travel phase and the restore reverses it.
-            # Stop and let the motor settle before commanding the opposite
-            # direction (issue #147) — an instant reversal leaves a short
-            # restore pulse's stop command dropped by the relay, so the cover
-            # overruns to its physical endpoint.
-            await self._async_handle_command(SERVICE_STOP_COVER)
-            if self._tilt_restore_superseded(epoch):
-                # Cancelled while stopping — bail before the settle delay so we
-                # don't block the background task for a dead restore.
-                self._log("_start_tilt_restore :: cancelled before settle delay")
-                return
-            await self._direction_change_delay()
-            if self._tilt_restore_superseded(epoch):
-                self._log("_start_tilt_restore :: cancelled during settle delay")
-                return
-            command = self._tilt_strategy.tilt_command_for(closing)
-            await self._async_handle_command(command)
-
-        if self._tilt_restore_superseded(epoch):
-            # The motor is already energized: whoever cancelled us sent their
-            # stop while we were awaiting this turn-on, so theirs went out
-            # first and ours landed after it. Nothing else will take it down —
-            # we are the only one that knows it went up — and on a latching
-            # relay that means a cover driving to its endpoint untracked.
-            self._log("_start_tilt_restore :: cancelled during motor start, stopping")
-            await self._stop_restore_motor()
-            return
-        self.tilt_calc.start_travel(restore_target)
-        self.start_auto_updater()
-
-    async def _maybe_start_recalibrated_leg(self) -> None:
-        """Run the second leg of a recalibrated move, if one is armed (#179).
-
-        Leg A drove the cover to a physical endpoint, so the tracker is now
-        true. Wait out any endpoint run-on and the motor settle gap, then make
-        the move the user actually asked for.
-        """
-        target = self._pending_recalibrated_target
-        if target is None:
-            return
-        axis = self._pending_recalibrated_axis
-        armed_epoch = self._recalibration_epoch
-        # Clear unconditionally, epoch match or not: a stale pending target
-        # left behind by a superseded leg A must not linger forever waiting
-        # for an epoch that will never come again.
-        self._disarm_recalibrated_leg()
-        if self._movement_epoch != armed_epoch:
-            return
-
-        # Leg A may have armed a delayed relay stop (endpoint_runon_time).
-        # Settling while the relay is still energized would make the rest gap a
-        # fiction, so let the run-on finish before starting the clock.
-        #
-        # asyncio.wait, not a bare `await delay_task`: a bare await makes
-        # delay_task this coroutine's own suspension point, so cancelling
-        # *this* task (e.g. the per-tick task from hass.async_create_task,
-        # torn down on entity removal or HA shutdown) would cancel delay_task
-        # right along with it -- killing the pending _delayed_stop before it
-        # de-energises the relay, leaving it latched on. asyncio.wait's own
-        # internal waiter absorbs that instead: delay_task keeps running to
-        # completion regardless, while a cancellation of this task still
-        # propagates out normally through the await.
-        delay_task = self._delay_task
-        if delay_task is not None and not delay_task.done():
-            await asyncio.wait({delay_task})
-
-        # The run-on wait above is itself an await — a stop or a new command
-        # can land while it is in flight, and would otherwise go unnoticed:
-        # _settle_before_reversing captures self._movement_epoch on entry and
-        # compares it to itself after its OWN sleep, so it only ever catches a
-        # supersede landing during its own wait — never one that already
-        # landed during the run-on wait above, before settle even started
-        # (issue #179). Re-check explicitly.
-        if self._movement_epoch != armed_epoch:
-            self._log("_maybe_start_recalibrated_leg :: superseded during run-on wait")
-            return
-
-        # auto_stop_if_necessary cleared _last_command, so leg B does not read
-        # itself as a direction change and would skip DIRECTION_CHANGE_DELAY --
-        # reversing the motor with no rest. Settle explicitly.
-        if not await self._settle_before_reversing():
-            self._log("_maybe_start_recalibrated_leg :: superseded during settle")
-            return
-
-        self._log("_maybe_start_recalibrated_leg :: %s leg to %d%%", axis, target)
-        try:
-            if axis == "tilt":
-                await self.set_tilt_position(target, recalibrate=False)
-            else:
-                await self.set_position(target, recalibrate=False)
-        except HomeAssistantError as err:
-            # This runs on a fresh per-tick task from hass.async_create_task,
-            # not literally "the auto-updater" — stop_auto_updater() has
-            # already unsubscribed the interval timer by the time we get
-            # here, so an escape would not take that down. It would still
-            # surface only as a noisy unhandled-task-exception log, and worse,
-            # silently abandon the move the user asked for. The cover is
-            # parked at a true endpoint (leg A's), so leaving it there is
-            # safe and correctly tracked — warn instead of losing it silently.
-            _LOGGER.warning(
-                "(%s) recalibrated move to %d%% failed: %s",
-                self.entity_id,
-                target,
-                err,
-            )
-
-    async def _stop_restore_motor(self) -> None:
-        """Take down whichever motor the restore energized.
-
-        Only when nothing is tracking it. Superseders come in two kinds: a
-        stop, which leaves no movement behind and so leaves our turn-on running
-        with nobody to switch it off — the case this exists for — and a new
-        *movement*, which abandoned the lifecycle and then energized a motor of
-        its own. Stopping in the second case kills the replacement's motor
-        while its calculator animates on, which is the same tracker-
-        confidently-wrong desync approached from the other side. A live
-        calculator is exactly the difference between the two.
-
-        _tilt_settle rather than a bare stop because at 0%/100% the motor has
-        already stopped on its own limit switch and re-pulsing a momentary
-        relay there would start it moving again. A path added to avoid leaving
-        a motor running must not do that either.
-        """
-        if self.travel_calc.is_traveling() or (
-            self._has_tilt_support() and self.tilt_calc.is_traveling()
-        ):
-            self._log("_stop_restore_motor :: a movement is tracking, leaving it alone")
-            return
-        if self._has_tilt_motor():
-            await self._tilt_settle()
-            self._on_tilt_motor_move_complete()
-        else:
-            await self._async_handle_command(SERVICE_STOP_COVER)
 
     # -----------------------------------------------------------------------
     # Relay command dispatch
